Added canvas angle to the GUI. Added a prompt to confirm changes made on pressing enter or escape after changes were made (i.e. undo stack isn't empty); direct clicks don't confirm. Fixed tab index flow.

+++ Logic/Setting.cs
@@ -82,7 +82,7 @@ namespace BrushFactory
                 { ShortcutTarget.ResetCanvasTransforms, new Setting(Localization.Strings.ResetCanvas, ShortcutTargetDataType.Action) },
                 { ShortcutTarget.CanvasX, new Setting(Localization.Strings.CanvasX, int.MinValue, int.MaxValue) },
                 { ShortcutTarget.CanvasY, new Setting(Localization.Strings.CanvasY, int.MinValue, int.MaxValue) },
-                { ShortcutTarget.CanvasRotation, new Setting(Localization.Strings.CanvasRotation, float.MinValue, float.MaxValue) },
+                { ShortcutTarget.CanvasRotation, new Setting(Localization.Strings.CanvasRotation, int.MinValue, int.MaxValue) },
                 { ShortcutTarget.BlendMode, new Setting(Localization.Strings.ShortcutBlendMode, 0, Enum.GetValues(typeof(BlendMode)).Length - 1) }
             };
         }
